+++ CHANGELOG.md
@@ -20,6 +20,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added endpoint `set_cooling_curve_slope()` to heating circuit
 - Added endpoint `get_cooling_curve_offset()` to heating circuit
 - Added endpoint `set_cooling_curve_offset()` to heating circuit
+- Added endpoint `get_target_cooling_temperature_day()` to heating circuit
+- Added endpoint `set_target_cooling_temperature_day()` to heating circuit
+- Added endpoint `set_heating_limit_day()` to heating circuit
+- Added endpoint `get_cooling_limit_day()` to heating circuit
+- Added endpoint `set_cooling_limit_day()` to heating circuit
+- Added endpoint `get_target_cooling_temperature_night()` to heating circuit
+- Added endpoint `set_target_cooling_temperature_night()` to heating circuit
+- Added endpoint `set_heating_limit_night()` to heating circuit
+- Added endpoint `get_cooling_limit_night()` to heating circuit
+- Added endpoint `get_cooling_limit_night()` to heating circuit
 
 ## [2.10.2] - 2026-04-03